+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - `ForEachCause` function to be able to loop through all errors in a chain
+- `Context` function to get the context from an error (and all parent errors)
 
 ### Changed